requirements.txt
pandas
requests
scikit-learn
streamlit
plotly
python-dotenv
numpy
matplotlib
